Implement ability to load sudoku puzzle from rest request Create debug rest request scene for testing ping and sudoku test capabilities against Python Flask api server
