+++ Prototype/Unity/Superdoku/Assets/Scenes/Debug.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f32b96c33b4caaa44b0f7ed6ed7a7c90
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
