+++ Prototype/Unity/Superdoku/Assets/Scripts/Utils/Utils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ecdbccbdf42062f4981da86aaa05a0cd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
